Write Oodle packets one at a time
Full disclaimer that this bugfix is very self-centered, we discovered
our networking code in Kawari is way too simple and doesn't handle
partial packets.

The previous behavior wrote the packet header and its compressed data
separately, but the client never seems to do this in practice. I rewrote
the code a bit so the newly compressed Oodle data is copied back into
the write buffer so it can be sent all as one thing.

I tested this against Kawari (which can now be captured) and retail.